save verifcation'

+++ setup.py
@@ -9,25 +9,25 @@ def start_wizard():
     system('clear')
     print('Welcome to Setup Wizard')
 
-    get_command()
-
 def get_command():
     # get the input from user and decide what to do with it. If command does not exist, function gets called again
     
     comm = input("Enter a command to get started: ")
 
     if comm == 'help':
-        system('cat helpfile')
-        get_command()
+        system('cat support/helpfile')
     elif comm == "add all":
         add_all()
     elif comm == 'add at':
         print('add at')
     elif comm in done_keywords:
         print('Goodbye.')
+        return False
     else: 
         print('[ERROR]: Invalid command. Please try again. Pass help for all commands')
-        get_command()
+    
+    return True
+        
 
 def verify_airport_is_valid() -> (str, str):
     # verifies ICAO ident provided is a valid ident. If not, user will be asked to provide a valid one before going on.
@@ -54,7 +54,7 @@ def add_all():
         airport = verify_airport_is_valid()
 
         if airport[0] == 'done':
-            save_file(config)
+            save_airports(config)
             break
         else:
             led = input("    LED position: ")
@@ -64,17 +64,25 @@ def add_all():
             print('[SUCCESS]: Assigned %s to LED at positon %s' % (airport[1], led))
             
 
-def save_file(configuration):
+def save_airports(configuration):
     config =  ConfigParser()
     config.read('raspi-metar.conf')
     config['airports'] = configuration
     
-    with open('raspi-metar.conf', 'w') as configfile:
-        config.write(configfile)
-    
-    print("[SUCCESS]: Saved airports to file.")
+    verify = input("Are you sure you want to save?(y/n) ")
 
-    get_command()
+    if verify == 'yes' or verify.lower() == 'y':
+        with open('raspi-metar.conf', 'w') as configfile:
+            config.write(configfile)
+        
+        print("[SUCCESS]: Saved airports to file.")
+    else:
+        print("Ok. Not saving recent configuration")
 
 
-start_wizard()
+if __name__ == "__main__":
+    start_wizard()
+
+    loop = True
+    while loop:
+        loop = get_command()
